refactor: types 분리
